+++ src/test/promisifyEvent.js
@@ -5,26 +5,47 @@ const { AssertionError } = require('assert');
 /**
  * Promisifies an event.
  * @param {object} emitter An event emitter with `on` and `off` methods.
- * @param {string} event The event name.
- * @param {number} [timeout=1000] How many milliseconds to wait for the event.
- * @returns {Promise<*>} Event data.
+ * @param {string} eventName The event name.
+ * @param {number} [msTimeLimit=1000] How many milliseconds to wait for the event.
+ * @returns {Promise<*>} Resolves the event data.
  * @see [Stack Overflow answer](https://stackoverflow.com/a/40353376/1596978).
  * @ignore
  */
-module.exports = function promisifyEvent(emitter, event, timeout = 1000) {
+module.exports = function promisifyEvent(
+  emitter,
+  eventName,
+  msTimeLimit = 1000
+) {
+  if (
+    typeof emitter !== 'object' ||
+    typeof emitter.on !== 'function' ||
+    typeof emitter.off !== 'function'
+  )
+    throw new TypeError(
+      'First argument `emitter` must be an event emitter with `on` and `off` methods.'
+    );
+
+  if (typeof eventName !== 'string')
+    throw new TypeError('Second argument `eventName` must be a string.');
+
+  if (typeof msTimeLimit !== 'number' || msTimeLimit < 0)
+    throw new TypeError(
+      'Third argument `msTimeLimit` must be a positive number.'
+    );
+
   return new Promise((resolve, reject) => {
     // Ensure the timeout error stack trace starts at the location where
     // `promisifyEvent` is called. Creating the error within the `setTimeout`
     // callback function would result in an unhelpful stack trace.
     const timeoutError = new AssertionError({
-      message: `Event “${event}” ${timeout} millisecond timeout.`,
+      message: `Event “${eventName}” ${msTimeLimit} millisecond timeout.`,
       stackStartFn: promisifyEvent,
     });
 
     const timer = setTimeout(() => {
-      emitter.off(event, listener);
+      emitter.off(eventName, listener);
       reject(timeoutError);
-    }, timeout);
+    }, msTimeLimit);
 
     /**
      * Listener for the event.
@@ -33,10 +54,10 @@ module.exports = function promisifyEvent(emitter, event, timeout = 1000) {
      */
     function listener(data) {
       clearTimeout(timer);
-      emitter.off(event, listener);
+      emitter.off(eventName, listener);
       resolve(data);
     }
 
-    emitter.on(event, listener);
+    emitter.on(eventName, listener);
   });
 };